backend/requirements.txt
# Backend API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
google-generativeai==0.3.1
# faiss-cpu==1.7.4  # [DEPRECATED] 已改用 PostgreSQL + pgvector
pydantic==2.5.0
python-multipart==0.0.6
pyyaml==6.0.1  # MCP 配置檔案支援
# date-extractor==0.7.0  # 暫時註解：網路問題時無法安裝，前兩個優先級（相對日期和 MMDD）已足夠

# Database dependencies
sqlalchemy==2.0.23  # SQLAlchemy ORM for PostgreSQL
psycopg2-binary==2.9.9  # PostgreSQL adapter

# PostgreSQL + pgvector dependencies
pgvector==0.3.0  # pgvector extension for PostgreSQL

# Embedding generation
# 注意：sentence-transformers 2.2.2 需要較舊版本的 huggingface-hub
# 如果遇到兼容性問題，請使用：sentence-transformers==2.7.0 或更新版本
sentence-transformers==2.7.0  # For embedding generation (paraphrase-multilingual-MiniLM-L12-v2)

# ReID (Re-identification) models for object embedding
# 優先使用 torchreid（FastReID 風格），如果不可用則使用 timm 或 torchvision
# torchreid  # 可選：如果需要完整的 ReID 功能，請取消註解（可能需要額外配置）
timm>=0.9.0  # 主要方案：使用 timm 的 ResNet50（輸出 2048 維）
torch>=2.0.0  # PyTorch（如果尚未安裝）
torchvision>=0.15.0  # 備用方案：使用 torchvision ResNet50（輸出 2048 維）

# YOLO object detection
ultralytics>=8.0.0  # YOLO 物件偵測模型（用於影片分析）

# Moondream / Hugging Face 模型載入（device_map 需 accelerate）
transformers>=4.40.0
accelerate>=0.25.0
safetensors>=0.4.0
# Monitoring and Diagnostics
psutil==5.9.6
pynvml==11.5.0  # NVIDIA GPU monitoring

gdown